Correct regex (#56)
The current regex fails to work with single digit version numbers. The docker images is trying to download a version of 0.13.0< where the < is coming from the HTML.

This PR updates the regex to not do that anymore and switches to use grep as the inbuilt bash regex does not support PCRE so does not support using lookarounds. The new regex uses a lookaround.